Make Iter auto-calculating, eliminate CalculatingIter
It turns out that using a cheap iterator which only wraps objects makes
the stack overflow with a very basic query:

    today()
        .yearly(1)
        .unwrap()
        .every(::indicator::Day::Monday)
        .every(::indicator::Month::January)
        .collect::<Vec<_>>()

So here we cgo and optimize the Iter type to automatically calculate all
the things right away. This makes it a bit more expensive to call, of
course, but is better than overflowing.